connect ancillary chunks to one file
