+++ Project1/main.py
@@ -14,12 +14,12 @@ def choose_photo():
     chunk_processor.print_chunks_types()
     chunk_processor.IHDR_chunk_processor()
     chunk_processor.IDAT_chunk_processor()
-    chunk_processor.tIME_chunk_prcessor()
     chunk_processor.gAMA_chunk_processor()
     chunk_processor.cHRM_chunk_processor()
     chunk_processor.tEXt_chunk_processor()
     chunk_processor.iTXt_chunk_processor()
     chunk_processor.zTXt_chunk_processor()
+    chunk_processor.tIME_chunk_prcessor()
     chunk_processor.IEND_chunk_processor()
     # query = chunk_processor.return_chunks_names_query()
     display_photo(chunk_processor)
